LJF done but commented

+++ threadpool.h
@@ -2,6 +2,7 @@
 #define THREADPOOL_H
 #include <pthread.h>
 #include <stdbool.h>
+#include <sys/types.h>
 
 typedef void (*thread_func_t)(void *arg);
 
@@ -9,7 +10,7 @@ typedef struct ThreadPool_work_t {
     thread_func_t func;              // The function pointer
     void *arg;                       // The arguments for the function
     struct ThreadPool_work_t *next; // each work should know which work comes next
-    int size; // file size of the work
+    off_t size; // file size of the work
 } ThreadPool_work_t;
 
 typedef struct {
